@@ -1,4 +1,6 @@
 import * as React from 'react';
+import toPath from 'lodash.topath';
+import cloneDeep from 'lodash.clonedeep';
 
 /** @private is the given object/value a promise? */
 export function isPromise(value: any): boolean {
@@ -27,9 +29,8 @@ export function values<T>(obj: any): T[] {
   return vals;
 }
 
-/** 
- * @private Deeply get a value from an object via it's dot path. 
- * See https://github.com/developit/dlv/blob/master/index.js
+/**
+ * @private Deeply get a value from an object via it's path.
  */
 export function dlv(
   obj: any,
@@ -37,22 +38,22 @@ export function dlv(
   def?: any,
   p: number = 0
 ) {
-  key = (key as string).split ? (key as string).split('.') : key;
-  while (obj && p < key.length) {
-    obj = obj[key[p++]];
+  const path = toPath(key);
+  while (obj && p < path.length) {
+    obj = obj[path[p++]];
   }
   return obj === undefined ? def : obj;
 }
 
-/** 
- * @private Deeply set a value from in object via it's dot path. 
+/**
+ * @private Deeply set a value from in object via it's path.
  * See https://github.com/developit/linkstate
  */
 export function setDeep(path: string, value: any, obj: any): any {
   let res: any = {};
   let resVal: any = res;
   let i = 0;
-  let pathArray = path.replace(/\]/g, '').split(/\.|\[/);
+  let pathArray = toPath(path);
 
   for (; i < pathArray.length - 1; i++) {
     const currentPath: string = pathArray[i];
@@ -61,9 +62,7 @@ export function setDeep(path: string, value: any, obj: any): any {
     if (resVal[currentPath]) {
       resVal = resVal[currentPath];
     } else if (currentObj) {
-      resVal = resVal[currentPath] = Array.isArray(currentObj)
-        ? [...currentObj]
-        : { ...currentObj };
+      resVal = resVal[currentPath] = cloneDeep(currentObj);
     } else {
       const nextPath: string = pathArray[i + 1];
       resVal = resVal[currentPath] =
@@ -78,15 +77,12 @@ export function setDeep(path: string, value: any, obj: any): any {
 /** @private is the given object a Function? */
 export const isFunction = (obj: any) => 'function' === typeof obj;
 
-
 /** @private is the given object an Object? */
 export const isObject = (obj: any) => obj !== null && typeof obj === 'object';
 
-/**
- * @private is the given object an Integer? 
- * see https://stackoverflow.com/questions/10834796/validate-that-a-string-is-a-positive-integer
-*/
+/** @private is the given object an integer  */
 export const isInteger = (obj: any) => String(Math.floor(Number(obj))) === obj;
 
+/** @private Does a React component have exactly 0 children? */
 export const isEmptyChildren = (children: any) =>
   React.Children.count(children) === 0;
